test(model-sources): extract stub upstream helpers for the WP-C1 suites (#2123)
Move `_free_port`, `_create_model_source`, `_enable_api_key_auth` and the stub
OpenAI-compatible upstream runner out of `test_model_source_routing.py` into
`tests/integration/model_source_helpers.py` so the transport-hardening and
source-dispatch suites can drive the same stub without importing a test
module. The runner is exposed as the `stub_source_upstreams()` async context
manager (a fixture imported across modules trips ruff F811 on every test
parameter); each module wraps it in a local `source_upstream` fixture.
`start()` gains `handler_cancellation` / `shutdown_timeout` for handlers that
deliberately stall until the client leaves.
